@@ -1,3 +1,5 @@
 module github.com/creack/pty
 
-go 1.18
+go 1.21.5
+
+require golang.org/x/sys v0.13.0
