@@ -1,24 +1,28 @@
 package pty
 
-import "os"
+// Winsize describes the terminal size.
+type Winsize struct {
+	Rows uint16 // ws_row: Number of rows (in cells).
+	Cols uint16 // ws_col: Number of columns (in cells).
+	X    uint16 // ws_xpixel: Width in pixels.
+	Y    uint16 // ws_ypixel: Height in pixels.
+}
 
 // InheritSize applies the terminal size of pty to tty. This should be run
 // in a signal handler for syscall.SIGWINCH to automatically resize the tty when
 // the pty receives a window size change notification.
-func InheritSize(pty, tty *os.File) error {
+func InheritSize(pty Pty, tty Tty) error {
 	size, err := GetsizeFull(pty)
 	if err != nil {
 		return err
 	}
-	return Setsize(tty, size)
+
+  return Setsize(tty, size)
 }
 
 // Getsize returns the number of rows (lines) and cols (positions
 // in each line) in terminal t.
-func Getsize(t *os.File) (rows, cols int, err error) {
+func Getsize(t FdHolder) (rows, cols int, err error) {
 	ws, err := GetsizeFull(t)
-	if err != nil {
-		return 0, 0, err
-	}
-	return int(ws.Rows), int(ws.Cols), nil
+	return int(ws.Rows), int(ws.Cols), err
 }
